move reinvite user to negative as test data cannot be created for positive test
